+++ lib/src/constants/key_aliases.dart
@@ -6,6 +6,8 @@ const keySourceLanguage = 'sourceLanguage';
 const keyTargetLanguages = 'targetLanguages';
 const keyTerms = 'terms';
 const keyId = 'id';
+const keyDefaultToBeTranslatedFromLanguage =
+    'defaultToBeTranslatedFromLanguage';
 const keyRelativeSourcePath = 'relativeSourcePath';
 const keyDescription = 'description';
 const keyExampleValues = 'exampleValues';