+++ lib/vy_dart_meme.dart
@@ -2,5 +2,5 @@
 export 'src/constants/key_aliases.dart';
 export 'src/constants/utils.dart';
 
-export 'src/vy_meme.dart';
+export 'src/meme.dart';